pyproject.toml
[project]
name = "revela"
version = "0.1.0"
description = "Right-click any chart or table on the web to get instant AI insights, powered by Gemma on Cloud Run"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.6.0",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "google-auth>=2.23.0",
    "requests>=2.31.0",
    "watchdog>=6.0.0",
]
